Device/Device manager/Device registration (#5)
This commit adds devices (CUDA and CPU, but easily extensible) and a device manager. Each component creates Python and C++ objects, and a Cython object for interconnect them. Each scheduler in different layer holds a device manager in each language. This commit maps a device in order for each task. But all tasks are actually running on CPU as at this commit, the Parla does not support device contexts.