Add a mechansim handling task placement and resource requirements
This commit adds APIs of placements and resourced requirements for a single device, an architecture and a multi-device task. This commit includes a deprecated semantic of the API that automatically changes unsupported devices or architectures to CPU. Specifying those unsupported devices will not be allowed through further commits. This API supports only FIFO task mapping. This commit passed bench.py for performance benchmarking.