Reverting format for synthetic tests,  Adding multidevice busywait.
Finally merging this old thing in as it didn't seem to have any comments and I needed multidevice busywait. 

IMPORTANT: This reverts the change for the graph format making it backwards compatible with the old branches and unit tests. "{data_size : location}, " vs "{data_size, location} |". Graphs generated since march will need to be regenerated to support this difference. This is mostly a temporary inconvenience as the custom graph format will be deprecated by YAML in an upcoming update.